# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Law Firm Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get notifications for current user
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
    post:
      operationId: createNotification
      tags: [notifications]
      summary: Create a notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNotificationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    put:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Lawyer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        oab:
          type: string
        specialty:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        isAdmin:
          type: boolean
        activeCaseCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, oab]
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        recipientId:
          type: integer
        senderId:
          type: integer
        sender:
          $ref: "#/components/schemas/Lawyer"
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [case_assigned, case_updated, deadline_reminder, message, system]
        caseId:
          type: integer
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, recipientId, title, message, type, read]
    CreateNotificationBody:
      type: object
      properties:
        recipientId:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [case_assigned, case_updated, deadline_reminder, message, system]
        caseId:
          type: integer
      required: [recipientId, title, message, type]
...